requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
alembic==1.6.5
anyio==3.3.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopage==0.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
clang==5.0
click==8.0.1
cliff==3.9.0
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==6.4.1
cycler==0.10.0
debugpy==1.4.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
eli5==0.11.0
entrypoints==0.3
facenet-pytorch==2.5.2
ffmpeg==1.4
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.7.0
fugashi==1.1.0
future==0.18.2
gast==0.4.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.17
greenlet==1.1.1
grpcio==1.39.0
h5py==3.1.0
idna==3.2
imageio==2.9.0
ipadic==1.0.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
Janome==0.4.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lab==6.4
lime==0.2.0.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mlxtend==0.18.0
mpmath==1.2.1
mtcnn==0.1.1
multidict==5.1.0
multiprocess==0.70.12.2
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nltk==3.6.2
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
optuna==2.9.1
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.3.0
prettytable==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydotplus==2.0.2
Pygments==2.9.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytorch-lightning==1.2.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.7.1
signate==0.9.8
simplejson==3.17.3
six==1.15.0
skater==1.1.2b1
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
stevedore==3.4.0
sympy==1.8
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.8
timm==0.4.12
tokenizers==0.10.3
torch==1.9.0
torchinfo==1.5.3
torchmetrics==0.4.1
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
transformers==4.5.0
txt2tags==3.7
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
yarl==1.6.3
